Add MUE_COMPILE_USE_SYSTEM_OPUS cmake option
This creates a simple way for using the system library version of opus which solves the problem of at same time dynamically linking to system libopus.so through libMuseSamplerCore.so and statically linking to a different version in an elegant way. Using this option can prevent conflicts arising because of this leading to hard-to-diagnose bugs/crashes (e.g. #21838)

To make it simpler the SetupOpusLibrary.cmake always creates a (sometimes alias/imported) target called "opus" which should make it unnecessary to add the include directories with MODULE_INCLUDE when target_link_library is called with this target. With either of DMUE_COMPILE_USE_SYSTEM_OPUS options it doesn't work and can't find the header files